% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWDPA.R
\name{getWDPA}
\alias{getWDPA}
\title{Get protected area}
\usage{
getWDPA(wdpaid, api_key, with_geometry = TRUE, filename = NA)
}
\arguments{
\item{wdpaid}{WDPA ID to return data for}

\item{api_key}{ProtectedPlanet.net API key}

\item{with_geometry}{Boolean to return the geometry of the protected area}

\item{filename}{If \code{with_geometry} is \code{TRUE}, the filename to use when saving the geometry to a GeoJSON file. By default, the filename is the WDPA ID (\emph{e.g.} 10721.geojson)}
}
\description{
This function gets the information available for the protected area specified with a WDPA ID.
}
\examples{
\dontrun{
getWDPA(wdpaid = 10721, api_key='xxxxxxxxxxx', with_geometry = TRUE)
}
}

