% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWDPACountry.R
\name{getWDPACountry}
\alias{getWDPACountry}
\title{Get the data available for a country in the ProtectedPlanet.net WDPA database}
\usage{
getWDPACountry(country_iso, with_geometry = TRUE, filename = NA, api_key)
}
\arguments{
\item{country_iso}{The ISO code for the country to search for}

\item{with_geometry}{Boolean to return the geometry of the country}

\item{filename}{If \code{with_geometry} is \code{TRUE}, the filename to use when saving the geometry to a GeoJSON file. By default, the filename is the ISO (\emph{e.g.} USA.geojson)}

\item{api_key}{ProtectedPlanet.net API key}
}
\description{
This function returns the data for a country in the ProtectedPlanet database. The country is identified by its ISO code, which can be obtained using the function \code{getWDPACountries()}.
}
\examples{
\dontrun{
regions <- getWDPACountry(country_iso = 'USA', api_key='xxxxxxxxxxx')
}
}

