% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWDPACountryAreas.R
\name{getWDPACountryAreas}
\alias{getWDPACountryAreas}
\title{Get the protected areas in a country from the ProtectedPlanet.net WDPA database}
\usage{
getWDPACountryAreas(country_iso, filename = NA, api_key)
}
\arguments{
\item{country_iso}{The ISO code for the country to search for}

\item{filename}{Store the list of areas found in this file in CSV format}

\item{api_key}{ProtectedPlanet.net API key}
}
\description{
This function searches the ProtectedPlanet database for protected areas in a country. The country is identified by its ISO code, which can be obtained using the function \code{getWDPACountries()}.
}
\examples{
\dontrun{
regions <- getWDPACountryAreas(country_iso = 'USA', api_key='xxxxxxxxxxx')
}
}

