% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwdpa-package.R
\docType{package}
\name{rwdpa-package}
\alias{rwdpa}
\alias{rwdpa-package}
\title{World Database of Protected Areas}
\description{
World Database of Protected Areas
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

