% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchWDPA.R
\name{searchWDPA}
\alias{searchWDPA}
\title{Search protected areas}
\usage{
searchWDPA(marine = NA, country = NA, designation = NA,
  jurisdiction = NA, governance = NA, iucn_category = NA, filename = NA,
  api_key)
}
\arguments{
\item{marine}{If \code{TRUE}, returns marine protected areas. If \code{FALSE}, returns terrestrial protected areas. Default is to return both.}

\item{country}{Returns areas matching the ISO3 code of the country.}

\item{designation}{Returns areas designated with the designation id (integer).}

\item{jurisdiction}{Returns areas designated with the jurisdiction code (integer).}

\item{governance}{Returns areas designated with the governance code (integer).}

\item{iucn_category}{Returns areas designated with the IUCN category code (integer).}

\item{filename}{Filename to store the results}

\item{api_key}{ProtectedPlanet.net API key}
}
\description{
Search protected areas in the WDPA database that match a set of parameters.
}
\examples{
\dontrun{
searchWDPA(marine = TRUE, country = 'CUB', filename = 'CUB.csv', api_key='xxxxxxxxxxx')
}
}

