% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testWDPAkey.R
\name{testWDPAkey}
\alias{testWDPAkey}
\title{Test an API key}
\usage{
testWDPAkey(api_key)
}
\arguments{
\item{api_key}{The API key to test}
}
\description{
This function test if an API key is valid
}
\examples{
\dontrun{
testWDPAkey(api_key="xxxxxxxxxxx")
}
}

