% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdpa_api_pa.R
\name{wdpa_api_pa}
\alias{wdpa_api_pa}
\title{WDPA API: protected areas}
\usage{
wdpa_api_pa(wdpa_id = NULL, geojson = FALSE, page = 1, per_page = 25,
  token = NULL, ...)
}
\arguments{
\item{wdpa_id}{(character) WDPA protected area ID}

\item{geojson}{(logical) If \code{TRUE} returns the GeoJSON representation of the
geometry of the protected areas. Default: \code{FALSE}}

\item{page}{(integer) page to retrieve. Default: 1}

\item{per_page}{(character) results to retrieve per page. Default: 25}

\item{token}{(character) token. see \link{rwdpa} for help on tokens}

\item{...}{curl options passed on to \link[crul:HttpClient]{crul::HttpClient}}
}
\description{
WDPA API: protected areas
}
\details{
This function uses the Protected Planet API
}
\examples{
\dontrun{
wdpa_api_pa(per_page = 1)
(x <- wdpa_api_pa(per_page = 1, geojson = TRUE))
x$geojson
wdpa_api_pa(per_page = 3, page = 2)
wdpa_api_pa(wdpa_id = 322152)
(x <- wdpa_api_pa(wdpa_id = 322152, geojson = TRUE))
x$geojson
}
}
\references{
\url{https://api.protectedplanet.net/documentation}
}
