% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdpa_api_pa_search.R
\name{wdpa_api_pa_search}
\alias{wdpa_api_pa_search}
\title{WDPA API: protected areas search}
\usage{
wdpa_api_pa_search(marine = FALSE, country = NULL, designation = NULL,
  jurisdiciton = NULL, governance = NULL, iucn_category = NULL,
  geojson = FALSE, page = 1, per_page = 25, token = NULL, ...)
}
\arguments{
\item{marine}{(logical) If \code{TRUE} returns all marine protected areas,
paginated. If \code{FALSE} (default) returns all terrestrial protected areas,
paginated.}

\item{country}{(character) If set, returns all protected areas from the
country with the given ISO3, paginated. By default, this is unset:
protected areas from all countries are returned.}

\item{designation}{(integer) If set, returns all protected areas designated
with the given \code{id}, paginated. By default, this is unset: protected areas
with all designations are returned.}

\item{governance}{(integer) If set, returns all protected areas with the
given governance \code{id}, paginated. By default, this is unset: protected
areas with all governances are returned.}

\item{iucn_category}{(integer) If set, returns all protected areas with the
given IUCN category \code{id}, paginated. By default, this is unset: protected
areas with all IUCN categories are returned.}

\item{geojson}{(logical) If \code{TRUE} returns the GeoJSON representation
of the geometry of the protected areas. Default: \code{FALSE}}

\item{page}{(integer) Controls the returned page. Default: 1}

\item{per_page}{(integer) Controls how many protected areas are returned per
page. Default: 25. For performance reasons, the maximum value is 50.}

\item{token}{(character) token. see \link{rwdpa} for help on tokens}

\item{...}{curl options passed on to \link[crul:HttpClient]{crul::HttpClient}}

\item{jurisdiction}{(integer) If set, returns all protected areas with the
given jurisdiciton \code{id}, paginated. By default, this is unset: protected
areas with all jurisdictions are returned.}
}
\value{
a data.frame
}
\description{
WDPA API: protected areas search
}
\details{
This function uses the Protected Planet API
}
\examples{
\dontrun{
wdpa_api_pa_search()
wdpa_api_pa_search(per_page = 3)
wdpa_api_pa_search(country = "BEL")
wdpa_api_pa_search(iucn_category = 3)
}
}
\references{
\url{https://api.protectedplanet.net/documentation}
}
