% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdpa_areas.R
\name{wdpa_area}
\alias{wdpa_area}
\title{Query a single Protected Area}
\usage{
wdpa_area(key = NULL, wdpa_id, with_geometry = FALSE)
}
\arguments{
\item{key}{your API key as provided by Protected Planet}

\item{wdpa_id}{a numeric vector giving the ID of the protected area}

\item{with_geometry}{boolean if geometry should be returned (for the moment
GeoJSON return is not supported)}
}
\description{
Return the characteristics of given protected area from its ID
}

