% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdpa_countries.R
\name{wdpa_countries}
\alias{wdpa_countries}
\title{List of all protected areas by country}
\usage{
wdpa_countries(key = NULL, with_geometry = FALSE, page = 1,
  per_page = 25)
}
\arguments{
\item{key}{your API key as provided by Protected Planet}

\item{with_geometry}{boolean if geometry should be returned (for the moment
GeoJSON return is not supported)}

\item{page}{integer indicating which page you want to return}

\item{per_page}{integer indicating the number of results you want to return}
}
\description{
Return the list and the characteristics of protected areas in the world.
}

