% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdpa_countries.R
\name{wdpa_country}
\alias{wdpa_country}
\title{Get a country list of Protected Areas}
\usage{
wdpa_country(key = NULL, country_iso3, with_geometry = FALSE)
}
\arguments{
\item{key}{your API key as provided by Protected Planet}

\item{country_iso3}{a character vector giving the ISO3 code of a country}

\item{with_geometry}{boolean if geometry should be returned (for the moment
GeoJSON return is not supported)}
}
\description{
Returns a list of protected area in a given country
}

