% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdpa_fetch.R
\name{wdpa_fetch}
\alias{wdpa_fetch}
\title{Fetch WDPA shp file}
\usage{
wdpa_fetch(path = "~/.rwdpa", overwrite = TRUE)
}
\arguments{
\item{path}{Base path to store file at. Default: \code{~/.rwdpa}}

\item{overwrite}{Overwrite if exists. Default: \code{TRUE}}
}
\description{
Fetch WDPA shp file
}
\details{
If  you don't have the metadata file, we download it for you
and read it in. After that we jus read the cached version. If for some
reason that file is deleted, we'll detect it and download again.
}
\examples{
\dontrun{
wdpa_fetch()
}
}

