% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdpa_meta.R
\name{wdpa_meta}
\alias{wdpa_meta}
\title{Fetch WDPA metadata}
\usage{
wdpa_meta(path = "~/.rwdpa", overwrite = TRUE)
}
\arguments{
\item{path}{Base path to store file at. Default: \code{~/.rwdpa}}

\item{overwrite}{Overwrite if exists. Default: \code{TRUE}}
}
\description{
Fetch WDPA metadata
}
\details{
If  you don't have the metadata file, we download it for you
and read it in. After that we jus read the cached version. If for some
reason that file is deleted, we'll detect it and download again.
}
\examples{
\dontrun{
wdpa_meta()
}
}
