% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdpa_meta_sites.R
\name{wdpa_meta_sites}
\alias{wdpa_meta_sites}
\title{Get WDPA metadata}
\usage{
wdpa_meta_sites(id = NULL, lat = NULL, long = NULL, type = NULL,
  page = NULL, ...)
}
\arguments{
\item{id}{(integer) Site id}

\item{lat}{(integer) latitude}

\item{long}{(integer) longitude}

\item{type}{(character) one of 'official' or 'public'. Default: both returned}

\item{page}{(integer) page to return, Default: 1}

\item{...}{Further named parameters passed on to \link[crul:HttpClient]{crul::HttpClient}}
}
\description{
Get WDPA metadata
}
\details{
This function uses the Protected Planet API
}
\examples{
\dontrun{
# wdpa_meta_sites(id = 40597)
# wdpa_meta_sites(lat = 51, long = 2.5)
# wdpa_meta_sites(page = 2)
# wdpa_meta_sites(type = 'official')
# wdpa_meta_sites(type = 'public')
}
}
\references{
\url{http://www.protectedplanet.net/api2/sites}
}
\keyword{internal}
