% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdpa_areas.R
\name{wdpa_search}
\alias{wdpa_search}
\title{Search protected areas based on attributes}
\usage{
wdpa_search(key = NULL, marine = NULL, country = NULL,
  designation = NULL, jurisdiction = NULL, governance = NULL,
  iucn_category = NULL, page = 1, per_page = 25)
}
\arguments{
\item{key}{your API key as provided by Protected Planet}

\item{marine}{boolean if \code{TRUE} returns all marine protected areas,
if \code{FALSE} returns all the terrestrial ones, \code{NULL}
by default returning all types of protected areas}

\item{country}{3 characters string if set returns all protected areas from
the given countries, by default unset and return PAs from
all countries}

\item{designation}{If set, returns all protected areas designated with the
given `id`, paginated. By default, this is unset:
protected areas with all designations are returned.}

\item{jurisdiction}{If set, returns all protected areas with the given
jurisdiciton `id`, paginated. By default, this is unset:
protected areas with all jurisdictions are returned.}

\item{governance}{If set, returns all protected areas with the given
governance `id`, paginated. By default, this is unset:
protected areas with all governances are returned.}

\item{iucn_category}{If set, returns all protected areas with the given IUCN
category `id`, paginated. By default, this is unset:
protected areas with all IUCN categories are returned.}

\item{page}{Controls the returned page. Defaults to 1.}

\item{per_page}{Controls how many protected areas are returned per page.
Defaults to 25. For performance reasons, the maximum value
is 50.}
}
\description{
Search protected areas based on attributes
}

