% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdpa_test.R
\name{wdpa_test}
\alias{wdpa_test}
\title{Test if API key is valid}
\usage{
wdpa_test(key = NULL)
}
\arguments{
\item{key}{your API key as provided by Protected Planet}
}
\description{
Check if your provided API key works with protected planet. Returns
\code{TRUE} if key is valid, returns \code{FALSE} otherwise.
}

