% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwdpa-package.R
\docType{package}
\name{rwdpa-package}
\alias{rwdpa-package}
\alias{rwdpa}
\title{World Database of Protected Areas}
\description{
World Database of Protected Areas
}
\section{WDPA Web API}{

\itemize{
\item API Docs: \url{https://api.protectedplanet.net/}
\item Get an API key at \url{https://api.protectedplanet.net/request}
}
}

\section{Package API}{

\itemize{
\item \code{\link[=rw_cache]{rw_cache()}} - manage bulk cached files
\item \code{\link[=wdpa_api_countries]{wdpa_api_countries()}} - API, countries
\item \code{\link[=wdpa_api_pa_search]{wdpa_api_pa_search()}} - API, search protected areas
\item \code{\link[=wdpa_api_pa]{wdpa_api_pa()}} - API, protected areas
\item \code{\link[=wdpa_fetch]{wdpa_fetch()}} - fetch a file
}
}

\author{
Scott Chamberlain \email{myrmecocystus+r@gmail.com}
}
\keyword{package}
