% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rWDPA.R
\docType{package}
\name{rWDPA-package}
\alias{rWDPA}
\alias{rWDPA-package}
\title{Package to query the WDPA (protectedplanet.net) API}
\description{
This package searches the ProtectedPlanet.net WDPA API. 
You will need to request an API key from 
https://api.protectedplanet.net before using this package.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{http://github.com/ljvillanueva/rWDPA/issues}
}

}
\author{
\strong{Maintainer}: Luis J. Villanueva-Rivera \email{ljvillanueva@coquipr.com}

}

