% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdpa_meta.R
\name{wdpa_meta}
\alias{wdpa_meta}
\title{Fetch WDPA metadata}
\usage{
wdpa_meta(path = "~/.rwdpa", overwrite = TRUE)
}
\arguments{
\item{path}{Base path to store file at. Default: \code{~/.rwdpa}}

\item{overwrite}{Overwrite if exists. Default: \code{TRUE}}
}
\description{
Fetch WDPA metadata
}
\examples{
\dontrun{
df <- wdpa_meta()
}
}

